Move the driver's Impl into a private implementation header
The z3 idiom (smt_context and its topic files): one class, several
translation units. IncrementalSolverImpl.h -- private to
lib/Incremental, never installed -- now holds the Impl state machine,
the helper classes it is built from (the encoding epoch, the backend
factory, the fragment cache types, the conjunct splitter) and the
refinement adapter; IncrementalSolver.cpp keeps the exact-stack route
and the public methods. The helpers lose their anonymous namespace for
ordinary inline linkage, since a per-TU copy of a type that shapes
Impl's layout would be an ODR violation the moment a second translation
unit joins. No code changed beyond linkage.
